Add launch scripts for the 7-zip binaries
The 7z binary uses Windows style library detection, so it's recommended to
place the binary to the same directory with the libraries. This is why the
launch scripts are needed here. So, we're following the same approach used
by e.g. the current Debian/Ubuntu package.

On the other hand, 7zr and 7za are standalone binaries, which wouldn't need
such an approach. There the launch scripts are used just for being consistent
with the implementation.